% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabMurDelayed}
\alias{tabMurDelayed}
\title{tabMurDelayed: a SummarizedExperiment instance with DelayedArray assay reference}
\format{SummarizedExperiment}
\source{
assay data from https://github.com/czbiohub/tabula-muris-vignettes/tree/master/data (https://s3.amazonaws.com/czbiohub-tabula-muris/TM_droplet_mat.rds)
}
\usage{
tabMurDelayed
}
\description{
tabMurDelayed: a SummarizedExperiment instance with DelayedArray assay reference
}
\note{
The numerical content is provided as a DelayedArray instance, hard coded to a specific HDF Lab resource.
}
\examples{
tabMurDelayed
}
\keyword{datasets}
